Optimizations and minor bug fixes with submodules.
The shclasses submodule directory has been combined
into a single file shclasses. This is non-optimal when
editing the file as it is very long. Nevertheless, I
could not find a way to correctly import the namespaces from
each file into the other files. If someone knows how to fix this,
please do...
